gdb/testsuite: fix gdb.mi/mi-condbreak-throw.exp failure
In commit:

  commit 949b569beb36a8fa68a163416542c9dfdfcdbd35
  Date:   Tue Aug 8 10:45:20 2023 +0100

      gdb: MI stopped events when unwindonsignal is on

a new test, gdb.mi/mi-condbreak-throw.exp, was added.  Unfortunately,
this test would fail when using the native-gdbserver board (and other
similar boards).

The problem was that one of the expected output patterns included some
output from the inferior.  When using the native-gdbserver board, this
output is not printed to GDB's tty, but is instead printed to
gdbserver's tty, the result is that the expected output no longer
matches, and the test fails.

Additionally, as the output is actually from the C++ runtime, rather
than the test's source file, changes to the C++ runtime could cause
the output to change.

To solve both of these issues, in this commit, I'm removing the
reference to the inferior's output, and replacing it with '.*', which
will skip the output if it is present, but is equally happy if the
output is not present.

After this commit gdb.mi/mi-condbreak-throw.exp now passes on all
boards, including native-gdbserver.